Decreasing life insurance

This function evaluates the n-year term decreasing life insurance. Both actuarial value and stochastic random sample can be returned.

DAxn(actuarialtable, x, n, 
  i=actuarialtable@interest,m = 0,k=1, 
  type = "EV", power=1)

Arguments

actuarialtable

An actuarial table object.

x

Age of the insured.

n

Length of the insurance period.

i

Interest rate, when present it overrides the interest rate of the actuarial table object.

m

Deferring period, even fractional, assumed 1 whether missing.

k

Number of fractional payments per period. Assumed to be 1 whether missing.

type

A string, either "EV" for expected value of the actuarial present value (default) or "ST" for one stochastic realization of the underlying present value of benefits. Alternatively, one can use "expected" or "stochastic" respectively (can be abbreviated).

power

The power of the APV. Default is 1 (mean)

Details

Formulas of Bowes book have been implemented.

Value

A numeric value representing the expected value or the simulated value.

References

Actuarial Mathematics (Second Edition), 1997, by Bowers, N.L., Gerber, H.U., Hickman, J.C., Jones, D.A. and Nesbitt, C.J.

Author

Giorgio A. Spedicato

Note

Neither fractional payments nor stochastic calculations have been implemented yet.

Warning

The function is provided as is, without any guarantee regarding the accuracy of calculation. We disclaim any liability for eventual losses arising from direct or indirect use of this software.

See also

Examples

    #using SOA illustrative life tables
    data(soaLt)
    soa08Act=with(soaLt, new("actuarialtable",interest=0.06,
    x=x,lx=Ix,name="SOA2008"))
    #evaluate the value of a 10 years decreasing term life insurance for an aged 25
    DAxn(actuarialtable=soa08Act, x=25, n=10) 
#> [1] 0.06070191