Function to evaluate life insurance.

This function evaluates n - years term and whole life insurance.

Axn(actuarialtable, x, n, i=actuarialtable@interest, 
  m, k=1, type = "EV", power=1, ...)

Arguments

actuarialtable

An actuarial table object.

x

Age of the insured. (can be a vector).

n

Coverage period, if missing the insurance is considered whole life \(n=\omega-x-m\). (can be a vector).

i

Interest rate (overrides the interest rate slot in actuarialtable). (should be a scalar).

m

Deferring period, even fractional, if missing assumed to be 0. (can be a vector).

k

Number of periods per year at the end of which the capital is payable in case of insured event, default=1 (capital payable at the end of death year). (should be a scalar).

type

A string, either "EV" for expected value of the actuarial present value (default) or "ST" for one stochastic realization of the underlying present value of benefits. Alternatively, one can use "expected" or "stochastic" respectively (can be abbreviated).

power

The power of the APV. Default is 1 (mean)

...

Arguments to be passed to pxt().

Details

The variance calculation has not been implemented yet.

Value

A numeric value representing either the actuarial value of the coverage (when type="EV") or a number drawn from the underlying distribution of Axn.

References

Actuarial Mathematics (Second Edition), 1997, by Bowers, N.L., Gerber, H.U., Hickman, J.C., Jones, D.A. and Nesbitt, C.J.

Author

Giorgio A. Spedicato

Note

It is possible that value returned by stochastic simulation are biased. Successive releases of this software will analyze the issue with detail.

Warning

The function is provided as is, without any guarantee regarding the accuracy of calculation. We disclaim any liability for eventual losses arising from direct or indirect use of this software.

See also

Examples

  
    #assume SOA example life table to be load
    data(soaLt)
    soa08Act=with(soaLt, new("actuarialtable",interest=0.06,
    x=x,lx=Ix,name="SOA2008"))
    #evaluate the value of a 40 years term life insurance for an aged 25
    Axn(actuarialtable=soa08Act, x=25, n=40) 
#> [1] 0.0479709
    #check an relevant life contingencies relationship
    k=12
    i=0.06
    j=real2Nominal(i,k)
    Axn(soa08Act, 30,k=12)
#> [1] 0.1052722
    i/j*Axn(soa08Act, 30,k=1)
#> [1] 0.1052722