Increasing annuity.

This function evaluates non - stochastic increasing annuities.

increasingAnnuity(i, n, type = "immediate")

Arguments

i

A numeric value representing the interest rate.

n

The number of periods.

type

The Payment type, either "advance" for the annuity due (default) or "arrears" for the annuity immediate. Alternatively, one can use "due" or "immediate" respectively (can be abbreviated).

Details

An increasing annuity shows the following flow of payments: \(1,2,\ldots,n-1,n\)

Value

The value of the annuity.

References

Broverman, S.A., Mathematics of Investment and Credit (Fourth Edition), 2008, ACTEX Publications.

Author

Giorgio A. Spedicato

Note

This function calls internally presentValue function.

Warning

The function is provided as is, without any guarantee regarding the accuracy of calculation. We disclaim any liability for eventual losses arising from direct or indirect use of this software.

See also

Examples

  #the present value of 1,2,...,n-1, n sequence of payments, 
  #payable at the end of the period
  #for 10 periods is
  increasingAnnuity(i=0.03, n=10)
#> [1] 44.83899
  #assuming a 3% interest rate