Various demographic functions

Various demographic functions

Lxt(object, x, t = 1, fxt = 0.5)

Tx(object, x)

Arguments

object

a lifetable or actuarialtable object

x

age of the subject

t

duration of the calculation

fxt

correction constant, default 0.5

Value

A numeric value

Details

Tx il the sum of years lived since age x by the population of the life table, it is the sum of Lx. The function is provided as is, without any warranty regarding the accuracy of calculations. Use at own risk.

References

Actuarial Mathematics (Second Edition), 1997, by Bowers, N.L., Gerber, H.U., Hickman, J.C., Jones, D.A. and Nesbitt, C.J.

Author

Giorgio Alfredo Spedicato.

Examples

data(soaLt)
soa08Act=with(soaLt, new("actuarialtable",interest=0.06,
x=x,lx=Ix,name="SOA2008"))
Lxt(soa08Act, 67,10)
#> [1] 61131812
#assumes SOA example life table to be load
data(soaLt)
soa08Act=with(soaLt, new("actuarialtable",interest=0.06,x=x,lx=Ix,name="SOA2008"))
Tx(soa08Act, 67)
#> [1] 102198948