Annuity function

Function to calculate present value of annuities-certain.

annuity(i, n, m = 0, k = 1, type = "immediate")

Arguments

i

Effective interest rate expressed in decimal form. E.g. 0.03 means 3%. It can be a vector of interest rates of the same length of periods.

n

Periods for payments. If n = infinity then annuity returns the value of a perpetuity (either immediate or due).

m

Deferring period, whose default value is zero.

k

Yearly payments frequency. A payment of \(k^-1\) is supposed to be performed at the end of each year.

type

The Payment type, either "advance" for the annuity due (default) or "arrears" for the annuity immediate. Alternatively, one can use "due" or "immediate" respectively (can be abbreviated).

Details

This function calculates the present value of a stream of fixed payments separated by equal interval of time. Annuity immediate has the fist payment at time t = 0, while an annuity due has the first payment at time t = 1.

Value

A string, either "immediate" or "due".

References

Broverman, S.A., Mathematics of Investment and Credit (Fourth Edition), 2008, ACTEX Publications.

Author

Giorgio A. Spedicato

Note

The value returned by annuity function derives from direct calculation of the discounted cash flow and not from formulas, like \({a^{\left( m \right)}}_{\left. {\overline {\, n \,}}\! \right| } = \frac{{1 - {v^n}}}{{{i^{\left( m \right)}}}}\). When m is greater than 1, the payment per period is assumed to be \(\frac{1}{m}\).

See also

Examples

# The present value of 5 payments of 1000 at one year interval that begins
# now when the interest rate is 2.5% is
1000 * annuity(i = 0.025, n = 5, type = "due")
#> [1] 4761.974
# A man borrows a loan of 20,000 to purchase a car at
# a nominal annual rate of interest of 0.06. He will pay back the loan through monthly
# installments over 5 years, with the first installment to be made one month
# after the release of the loan. What is the monthly installment he needs to pay?
20000 / annuity(i = 0.06 / 12, n = 5 * 12)
#> [1] 386.656