Functions to evaluate survival, death probabilities and deaths.

These functions evaluate raw survival and death probabilities between age x and x+t

dxt(object, x, t, decrement)
pxt(object, x, t, fractional = "linear", decrement)
qxt(object, x, t, fractional = "linear", decrement)

Arguments

object

A lifetable object.

x

Age of life x. (can be a vector for pxt, qxt).

t

Period until which the age shall be evaluated. Default value is 1. (can be a vector for pxt, qxt).

fractional

Assumptions for fractional age. One of "linear", "hyperbolic", "constant force" (can be abbreviated).

decrement

The reason of decrement (only for mdt class objects). Can be either an ordinal number or the name of decrement

Details

Fractional assumptions are:

linear: linear interpolation between consecutive ages, i.e. assume uniform distribution.

constant force of mortality : constant force of mortality, also known as exponential interpolation.

hyperbolic: Balducci assumption, also known as harmonic interpolation.

Note that fractional="uniform", "exponential", "harmonic" or "Balducci" is also authorized. See references for details.

Value

A numeric value representing requested probability.

References

Actuarial Mathematics (Second Edition), 1997, by Bowers, N.L., Gerber, H.U., Hickman, J.C., Jones, D.A. and Nesbitt, C.J.

Author

Giorgio A. Spedicato

Note

Function dxt accepts also fractional value of t. Linear interpolation is used in such case. These functions are called by many other functions.

Warning

The function is provided as is, without any warranty regarding the accuracy of calculations. The author disclaims any liability for eventual losses arising from direct or indirect use of this software.

See also

Examples

  #dxt example
  data(soa08Act)
  dxt(object=soa08Act, x=90, t=2)
#> [1] 3757.835
  #qxt example
  qxt(object=soa08Act, x=90, t=2)
#> [1] 0.3550183
  #pxt example
  pxt(object=soa08Act, x=90, t=2, "constant force" )
#> [1] 0.6449817
  #add another example for MDT